Detached PortList, PortRange(Type) from GMP monolith
